+++ OFM.Infrastructure.WebAPI/Services/Processes/ID365ProcessProvider.cs
@@ -5,7 +5,7 @@ namespace OFM.Infrastructure.WebAPI.Services.Processes;
 
 public interface ID365ProcessProvider
 {
-    int ProcessId { get; }
+    Int16 ProcessId { get; }
     string ProcessName { get; }
     Task<ProcessData> GetData();
     Task<ProcessResult> RunProcessAsync(ID365AppUserService appUserService, ID365WebApiService d365WebApiService);
